requirements_App_02.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras